Use shared locations panel for office and tsp - general cleanup
